+++ src/general.cpp
@@ -4,15 +4,15 @@
 //////////////////////////////////////
 #include "../include/general.hpp"
 
-double xlinearInter(
-  const double x1,
-  const double y1,
-  const double x2,
-  const double y2,
-  const double y
+float xlinearInter(
+  const float x1,
+  const float y1,
+  const float x2,
+  const float y2,
+  const float y
 )
 {
-  double x = 0.0;
+  float x = 0.0;
 
   x = x1 + (y - y1)*(x2 - x1)/(y2 - y1);
